+++ packages/ai-openai/src/model-meta.ts
@@ -1698,7 +1698,7 @@ const GPT_4 = {
 } as const satisfies ModelMeta<
   OpenAIBaseOptions & OpenAIStreamingOptions & OpenAIMetadataOptions
 >
-/* 
+/*
 const GPT_4O_MINI_TRANSCRIBE = {
   name: 'gpt-4o-mini-transcribe',
   context_window: 16_000,
@@ -1767,7 +1767,7 @@ const GPT_4O_TRANSCRIBE = {
 } as const satisfies ModelMeta<
   OpenAIBaseOptions & OpenAIStreamingOptions & OpenAIMetadataOptions
 > */
-/* 
+/*
 const GPT_4O_TRANSCRIBE_DIARIZE = {
   name: 'gpt-4o-transcribe-diarize',
   context_window: 16_000,